+++ lib/online_migrations/schema_statements.rb
@@ -182,31 +182,16 @@ module OnlineMigrations
 
     # Same as `initialize_column_rename` but for multiple columns.
     #
-    # The technique is built on top of database views, using the following steps:
-    #   1. Rename the table to some temporary name
-    #   2. Create a VIEW using the old table name with addition of the new columns as an aliases of the old ones
-    #   3. Add a workaround for ActiveRecord's schema cache
+    # This is useful to avoid multiple iterations of the safe column rename steps
+    # when renaming multiple columns.
     #
     # @param table_name [String, Symbol] table name
     # @param old_new_column_hash [Hash] the hash of old and new columns
     #
-    # @return [void]
-    #
     # @example
-    #   initialize_columns_rename(:users, {fname: :first_name, lname: :last_name})
+    #   initialize_columns_rename(:users, { fname: :first_name, lname: :last_name })
     #
-    # @note
-    #   Prior to using this method, you need to register the database table so that
-    #   it instructs ActiveRecord to fetch the database table information (for SchemaCache)
-    #   using the original table name (if it's present). Otherwise, fall back to the old table name:
-    #
-    #   ```OnlineMigrations.config.column_renames[table_name] = { old_column_name => new_column_name }```
-    #
-    #   Deploy this change before proceeding with this helper.
-    #   This is necessary to avoid errors during a zero-downtime deployment.
-    #
-    # @note None of the DDL operations involving original table name can be performed
-    #   until `finalize_column_rename` is run
+    # @see #initialize_column_rename
     #
     def initialize_columns_rename(table_name, old_new_column_hash)
       transaction do
@@ -217,9 +202,9 @@ module OnlineMigrations
     # Reverts operations performed by initialize_column_rename
     #
     # @param table_name [String, Symbol] table name
-    # @param _column_name [String, Symbol] the name of the column to be renamed.
+    # @param column_name [String, Symbol] the name of the column to be renamed.
     #     Passing this argument will make this change reversible in migration
-    # @param _new_column_name [String, Symbol] new new name of the column.
+    # @param new_column_name [String, Symbol] new new name of the column.
     #     Passing this argument will make this change reversible in migration
     #
     # @return [void]
@@ -227,8 +212,8 @@ module OnlineMigrations
     # @example
     #   revert_initialize_column_rename(:users, :name, :first_name)
     #
-    def revert_initialize_column_rename(table_name, _column_name = nil, _new_column_name = nil)
-      revert_initialize_columns_rename(table_name, { _column_name => _new_column_name })
+    def revert_initialize_column_rename(table_name, column_name = nil, new_column_name = nil)
+      revert_initialize_columns_rename(table_name, { column_name => new_column_name })
     end
 
     # Same as `revert_initialize_column_rename` but for multiple columns.
@@ -240,7 +225,7 @@ module OnlineMigrations
     # @return [void]
     #
     # @example
-    #   revert_initialize_columns_rename(:users, {fname: :first_name, lname: :last_name})
+    #   revert_initialize_columns_rename(:users, { fname: :first_name, lname: :last_name })
     #
     def revert_initialize_columns_rename(table_name, _old_new_column_hash = nil)
       transaction do
@@ -267,14 +252,14 @@ module OnlineMigrations
     # @return [void]
     #
     # @example
-    #   finalize_columns_rename(:users, {fname: :first_name, lname: :last_name})
+    #   finalize_columns_rename(:users, { fname: :first_name, lname: :last_name })
     #
     def finalize_columns_rename(table_name, old_new_column_hash)
       transaction do
         execute("DROP VIEW #{quote_table_name(table_name)}")
         rename_table("#{table_name}_column_rename", table_name)
-        old_new_column_hash.each_key do |old|
-          rename_column(table_name, old, old_new_column_hash[old])
+        old_new_column_hash.each do |column_name, new_column_name|
+          rename_column(table_name, column_name, new_column_name)
         end
       end
     end
@@ -297,12 +282,12 @@ module OnlineMigrations
     # @return [void]
     #
     # @example
-    #   revert_finalize_columns_rename(:users, {fname: :first_name, lname: :last_name})
+    #   revert_finalize_columns_rename(:users, { fname: :first_name, lname: :last_name })
     #
     def revert_finalize_columns_rename(table_name, old_new_column_hash)
       transaction do
-        old_new_column_hash.each_key do |old|
-          rename_column(table_name, old_new_column_hash[old], old)
+        old_new_column_hash.each do |column_name, new_column_name|
+          rename_column(table_name, new_column_name, column_name)
         end
         rename_table_create_view(table_name, old_new_column_hash)
       end
@@ -1170,9 +1155,10 @@ module OnlineMigrations
       def rename_table_create_view(table_name, old_new_column_hash)
         tmp_table = "#{table_name}_column_rename"
         rename_table(table_name, tmp_table)
-        column_mapping = old_new_column_hash.keys.map do |old|
-          "#{quote_column_name(old)} AS #{quote_column_name(old_new_column_hash[old])}"
+        column_mapping = old_new_column_hash.map do |column_name, new_column_name|
+          "#{quote_column_name(column_name)} AS #{quote_column_name(new_column_name)}"
         end.join(", ")
+
         execute(<<-SQL.squish)
           CREATE VIEW #{quote_table_name(table_name)} AS
             SELECT *, #{column_mapping}